css/theme/markup cleanup
