+++ designer/client/windowManager/index.ts
@@ -3,3 +3,4 @@ export {contentGetter} from "./ContentGetter"
 export {PromptContent} from "./PromptContent"
 export {WindowContent} from "./WindowContent"
 export {WindowKind} from "./WindowKind"
+export {WindowManager} from "./WindowManager"
